requirements.txt
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
helium==3.0.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parsel==1.7.0
playwright==1.31.0
pluggy==1.0.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyOpenSSL==23.0.0
PySocks==1.7.1
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
pywin32==305
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
scrapy-playwright==0.0.26
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
text-unidecode==1.3
tldextract==3.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.5.0
urllib3==1.26.14
w3lib==2.1.1
wsproto==1.2.0
xlrd==2.0.1
xlwings==0.29.1
zope.interface==5.5.2
